[IMP] account_payment_group_financial_surcharge:
Now Surcharge Coefficient on Fincancial plans admid percentages with
decimals.

closes ingadhoc/account-payment#312

Ticket: 52240

+++ account_payment_group_financial_surcharge/models/account_financing_plan.py
@@ -15,7 +15,7 @@ class AccountFinancingPlan(models.Model):
         default=10,
     )
     name = fields.Char(required=True)
-    surcharge_coefficient = fields.Integer(
+    surcharge_coefficient = fields.Float(
         'Surcharge coefficient',
         required=True,
     )
